Step 1: config in its own dir
