+++ src/config/kube.rs
@@ -1,4 +1,4 @@
-// Copyright 2017 Databricks, Inc.
+ // Copyright 2017 Databricks, Inc.
 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,50 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Handle reading .kube/config files
+//! Code to represent the data found in .kube/config files after it's deserialized, validated, and
+//! so on.  Data in here is what gets passed around to the rest of Click.
 
-use atomicwrites::{AtomicFile, AllowOverwrite};
-use chrono::{DateTime, Local, TimeZone};
-use chrono::offset::Utc;
-use duct::cmd;
-use serde_json::{self, Value};
-use serde_yaml;
-use rustyline::config as rustyconfig;
-
-use std::cell::RefCell;
 use std::collections::HashMap;
 use std::convert::From;
 use std::env;
 use std::error::Error;
-use std::fmt;
 use std::fs::File;
-use std::io::{self, BufReader, Read};
+use std::io::{BufReader, Read};
 
 use error::{KubeErrNo, KubeError};
 use kube::{Kluster, KlusterAuth};
 use certs::{get_cert, get_cert_from_pem, get_key_from_str, get_private_key};
 
-/// Kubernetes cluster config
-#[derive(Debug, Deserialize)]
-struct IConfig {
-    clusters: Vec<ICluster>,
-    contexts: Vec<IContext>,
-    users: Vec<IUser>,
-}
-
-#[derive(Debug, Deserialize)]
-struct ICluster {
-    name: String,
-    #[serde(rename = "cluster")] conf: IClusterConf,
-}
-
-#[derive(Debug, Deserialize)]
-struct IClusterConf {
-    #[serde(rename = "certificate-authority")] pub cert: Option<String>,
-    #[serde(rename = "certificate-authority-data")] pub cert_data: Option<String>,
-    #[serde(rename = "insecure-skip-tls-verify")] pub skip_tls: Option<bool>,
-    pub server: String,
-}
+use super::kubefile::AuthProvider;
 
 #[derive(Debug)]
 pub struct ClusterConf {
@@ -82,184 +53,6 @@ impl ClusterConf {
     }
 }
 
-#[derive(Debug, Deserialize)]
-struct IContext {
-    name: String,
-    #[serde(rename = "context")] conf: ContextConf,
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct ContextConf {
-    pub cluster: String,
-    //#[serde(default = "default")]
-    pub namespace: Option<String>,
-    pub user: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct IUser {
-    name: String,
-    #[serde(rename = "user")] conf: IUserConf,
-}
-
-// Classes to hold deserialized data for auth
-#[derive(Debug, Deserialize, Clone)]
-pub struct AuthProvider {
-    name: String,
-    pub token: RefCell<Option<String>>,
-    pub expiry: RefCell<Option<String>>,
-    pub config: AuthProviderConfig,
-}
-
-impl AuthProvider {
-    // Copy the token and expiry out of the config into the refcells
-    fn copy_up(&self) {
-        let mut token = self.token.borrow_mut();
-        *token = self.config.access_token.clone();
-        let mut expiry = self.expiry.borrow_mut();
-        *expiry = self.config.expiry.clone();
-    }
-
-    // true if expiry is before now
-    fn check_dt<T: TimeZone>(&self, expiry: DateTime<T>) -> bool {
-        let etime = expiry.with_timezone(&Utc);
-        let now = Utc::now();
-        etime < now
-    }
-
-    fn is_expired(&self) -> bool {
-        let expiry = self.expiry.borrow();
-        match *expiry {
-            Some(ref e) => {
-                // Somehow google sometimes puts a date like "2018-03-31 22:22:01" in the config
-                // and other times like "2018-04-01T05:57:31Z", so we have to try both.  wtf google.
-                if let Ok(expiry) = DateTime::parse_from_rfc3339(e) {
-                    self.check_dt(expiry)
-                } else if let Ok(expiry) = Local.datetime_from_str(e, "%Y-%m-%d %H:%M:%S") {
-                    self.check_dt(expiry)
-                } else {
-                    true
-                }
-            }
-            None => {
-                println!("No expiry set, cannot validate if token is still valid");
-                false
-            }
-        }
-    }
-
-    // Turn a {.credential.expiry_key} type string into a serde_json pointer string like
-    // /credential/expiry_key
-    fn make_pointer(&self, s: &str) -> String {
-        let l = s.len() - 1;
-        let split = &s[1..l].split('.');
-        split.clone().collect::<Vec<&str>>().join("/")
-    }
-
-    fn update_token(&self, token: &mut Option<String>, expiry: &mut Option<String>) {
-        match self.config.cmd_path {
-            Some(ref conf_cmd) => {
-                let args = self.config
-                    .cmd_args
-                    .as_ref()
-                    .map(|argstr| argstr.split_whitespace().collect())
-                    .unwrap_or(vec![]);
-                match cmd(conf_cmd, &args).read() {
-                    Ok(output) => {
-                        let v: Value = serde_json::from_str(output.as_str()).unwrap();
-                        let mut updated_token = false;
-                        match self.config.token_key.as_ref() {
-                            Some(ref tk) => {
-                                let token_pntr = self.make_pointer(tk.as_str());
-                                let extracted_token =
-                                    v.pointer(token_pntr.as_str()).and_then(|tv| tv.as_str());
-                                *token = extracted_token.map(|t| t.to_owned());
-                                updated_token = true;
-                            }
-                            None => {
-                                println!("No token-key in auth-provider, cannot extract token");
-                            }
-                        }
-
-                        if updated_token {
-                            match self.config.expiry_key.as_ref() {
-                                Some(ref ek) => {
-                                    let expiry_pntr = self.make_pointer(ek.as_str());
-                                    let extracted_expiry =
-                                        v.pointer(expiry_pntr.as_str()).and_then(|ev| ev.as_str());
-                                    *expiry = extracted_expiry.map(|e| e.to_owned());
-                                }
-                                None => {
-                                    println!(
-                                        "No expiry-key in config, will have to pull a new \
-                                         token on every command"
-                                    );
-                                }
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        println!("Failed to run update command: {}", e);
-                    }
-                }
-            }
-            None => {
-                println!("No update command specified, can't update");
-            }
-        }
-    }
-
-    /// Checks that we have a valid token, and if not, attempts to update it based on the config
-    pub fn ensure_token(&self) -> String {
-        let mut token = self.token.borrow_mut();
-        if token.is_none() || self.is_expired() {
-            // update
-            let mut expiry = self.expiry.borrow_mut();
-            *token = None;
-            self.update_token(&mut token, &mut expiry)
-        }
-        match token.as_ref() {
-            Some(t) => t.clone(),
-            None => {
-                println!(
-                    "Couldn't get an authentication token. You can try exiting Click and \
-                     running a kubectl command against the cluster to refresh it. Also please \
-                     report this error on the Click github page."
-                );
-                "".to_owned()
-            }
-        }
-    }
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct AuthProviderConfig {
-    #[serde(rename = "access-token")] pub access_token: Option<String>,
-    expiry: Option<String>,
-
-    #[serde(rename = "cmd-args")] cmd_args: Option<String>,
-    #[serde(rename = "cmd-path")] cmd_path: Option<String>,
-    #[serde(rename = "expiry-key")] expiry_key: Option<String>,
-    #[serde(rename = "token-key")] token_key: Option<String>,
-}
-
-/// This represents what we can find in a user in the actual config file (note the Deserialize).
-/// Hence all the optional fields.  At some point we should write a custom deserializer for this to
-/// make it cleaner
-#[derive(Debug, Deserialize, Clone)]
-pub struct IUserConf {
-    pub token: Option<String>,
-
-    #[serde(rename = "client-certificate")] pub client_cert: Option<String>,
-    #[serde(rename = "client-key")] pub client_key: Option<String>,
-    #[serde(rename = "client-certificate-data")] pub client_cert_data: Option<String>,
-    #[serde(rename = "client-key-data")] pub client_key_data: Option<String>,
-
-    pub username: Option<String>,
-    pub password: Option<String>,
-
-    #[serde(rename = "auth-provider")] pub auth_provider: Option<AuthProvider>,
-}
 
 // These are the different kinds of authentication data  a user might have
 
@@ -275,8 +68,8 @@ pub enum UserConf {
     Unsupported,
 }
 
-impl From<IUserConf> for UserConf {
-    fn from(iconf: IUserConf) -> UserConf {
+impl From<super::kubefile::UserConf> for UserConf {
+    fn from(iconf: super::kubefile::UserConf) -> UserConf {
         if let Some(token) = iconf.token {
             UserConf::Token(token)
         } else if let (Some(username), Some(password)) = (iconf.username, iconf.password) {
@@ -297,17 +90,6 @@ impl From<IUserConf> for UserConf {
     }
 }
 
-impl IConfig {
-    fn from_file(path: &str) -> Result<IConfig, io::Error> {
-        let f = File::open(path)?;
-        serde_yaml::from_reader(f).map_err(|e| {
-            io::Error::new(
-                io::ErrorKind::Other,
-                format!("Couldn't read yaml in '{}': {}", path, e.description()),
-            )
-        })
-    }
-}
 
 /// A kubernetes config
 // This is actual config we expose
@@ -315,7 +97,7 @@ impl IConfig {
 pub struct Config {
     pub source_file: String,
     pub clusters: HashMap<String, ClusterConf>,
-    pub contexts: HashMap<String, ContextConf>,
+    pub contexts: HashMap<String, super::kubefile::ContextConf>,
     pub users: HashMap<String, UserConf>,
 }
 
@@ -419,7 +201,7 @@ impl Config {
     pub fn from_files(paths: &[String]) -> Result<Config, KubeError> {
         let iconfs = paths
             .into_iter()
-            .map(|config_path| IConfig::from_file(config_path))
+            .map(|config_path| super::kubefile::Config::from_file(config_path))
             .collect::<Result<Vec<_>,_>>()?;
 
         // copy over clusters
@@ -580,121 +362,3 @@ impl Config {
     }
 }
 
-/// Click config
-#[derive(Debug, Deserialize, Serialize)]
-pub struct Alias {
-    pub alias: String,
-    pub expanded: String,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub enum EditMode {
-    Emacs,
-    Vi,
-}
-
-impl Default for EditMode {
-    fn default() -> Self { EditMode::Emacs }
-}
-
-impl fmt::Display for EditMode {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", match self {
-            EditMode::Emacs => "Emacs",
-            EditMode::Vi => "Vi",
-        })
-    }
-}
-
-impl Into<String> for &EditMode {
-    fn into(self) -> String {
-        format!("{}", self)
-    }
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub enum CompletionType {
-    Circular,
-    List,
-}
-
-impl Default for CompletionType {
-    fn default() -> Self { CompletionType::Circular }
-}
-
-impl fmt::Display for CompletionType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", match self {
-            CompletionType::Circular => "Circular",
-            CompletionType::List => "List",
-        })
-    }
-}
-
-impl Into<String> for &CompletionType {
-    fn into(self) -> String {
-        format!("{}", self)
-    }
-}
-
-#[derive(Debug, Default, Deserialize, Serialize)]
-pub struct ClickConfig {
-    pub namespace: Option<String>,
-    pub context: Option<String>,
-    pub editor: Option<String>,
-    pub terminal: Option<String>,
-    #[serde(default = "EditMode::default")]
-    pub editmode: EditMode,
-    #[serde(default = "CompletionType::default")]
-    pub completiontype: CompletionType,
-    #[serde(default = "Vec::new")]
-    pub aliases: Vec<Alias>,
-}
-
-impl ClickConfig {
-    pub fn from_file(path: &str) -> ClickConfig {
-        match File::open(path) {
-            Ok(f) => match serde_yaml::from_reader(f) {
-                Ok(c) => c,
-                Err(e) => {
-                    println!("Could not read config file {:?}, using default values", e);
-                    ClickConfig::default()
-                }
-            },
-            Err(e) => {
-                println!(
-                    "Could not open config file at '{}': {}. Using default values",
-                    path, e
-                );
-                ClickConfig::default()
-            }
-        }
-    }
-
-    pub fn get_rustyline_conf(&self) -> rustyconfig::Config {
-        let mut config = rustyconfig::Builder::new();
-        config = match self.editmode {
-            EditMode::Emacs => config.edit_mode(rustyconfig::EditMode::Emacs),
-            EditMode::Vi => config.edit_mode(rustyconfig::EditMode::Vi),
-        };
-        config = match self.completiontype {
-            CompletionType::Circular =>
-                config.completion_type(rustyconfig::CompletionType::Circular),
-            CompletionType::List =>
-                config.completion_type(rustyconfig::CompletionType::List),
-        };
-        config.build()
-    }
-
-    /// Save this config to specified path.  It's safe to call this from multiple running instances
-    /// of Click, since we use an AtomicFile
-    pub fn save_to_file(&self, path: &str) -> Result<(), KubeError> {
-        let af = AtomicFile::new(path, AllowOverwrite);
-        try!(af.write(|mut f| {
-            serde_yaml::to_writer(&mut f, &self)
-        }).map_err(|e| KubeError::ConfigFileError(
-            format!("Failed to write config file: {}", e.description())
-        )));
-        Ok(())
-    }
-}
