+++ src/config/mod.rs
@@ -1,0 +1,14 @@
+mod click;
+mod kube;
+mod kubefile;
+
+
+pub use self::click::Alias;
+pub use self::click::CompletionType;
+pub use self::click::ClickConfig;
+pub use self::click::EditMode;
+
+pub use self::kube::Config;
+
+pub use self::kubefile::AuthProvider;
+pub use self::kubefile::ContextConf;
